src/backend/pyproject.toml
#
# find package
#
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "find"
version = "0.0.1"
authors = [{ "name" = "DINUM", "email" = "dev@mail.numerique.gouv.fr" }]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Django",
  "Framework :: Django :: 6",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]
description = "An application to print markdown to pdf from a set of managed templates."
keywords = ["Django", "Contacts", "Templates", "RBAC"]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
  "celery[redis]==5.6.3",
  "django-bolt==0.8.0",
  "django-configurations==2.5.1",
  "django-cors-headers==4.9.0",
  "django-lasuite[all]==0.0.26",
  "django-redis==6.0.0",
  "django==6.0.5",
  "dockerflow==2026.3.4",
  "gunicorn==25.3.0",
  "mozilla-django-oidc==5.0.2",
  "msgspec==0.21.1",
  "opensearch-py==3.2.0",
  "psycopg[binary]==3.3.4",
  "py3langid==0.3.0",
  "pyjwt==2.12.1",
  "redis==6.4.0",
  "requests==2.33.1",
  "sentry-sdk==2.58.0",
  "whitenoise==6.12.0",
]

[project.urls]
"Bug Tracker" = "https://github.com/suitenumerique/find/issues/new"
"Changelog" = "https://github.com/suitenumerique/find/blob/main/CHANGELOG.md"
"Homepage" = "https://github.com/suitenumerique/find"
"Repository" = "https://github.com/suitenumerique/find"

[project.optional-dependencies]
dev = [
  "django-extensions==4.1",
  "factory_boy==3.3.3",
  "faker==40.15.0",
  "ipdb==0.13.13",
  "ipython==9.13.0",
  "pyfakefs==6.2.0",
  "pylint-django==2.7.0",
  "pylint==4.0.5",
  "pytest-cov==7.1.0",
  "pytest-django==4.12.0",
  "pytest-icdiff==0.9",
  "pytest-rerunfailures==15.0.0",
  "pytest-xdist==3.8.0",
  "pytest==9.0.3",
  "responses==0.26.0",
  "ruff==0.15.12",
  "types-requests==2.33.0.20260503",
]

[tool.setuptools]
packages = { find = { where = ["."], exclude = ["tests"] } }
zip-safe = true

[tool.distutils.bdist_wheel]
universal = true

[tool.ruff]
exclude = [
  ".git",
  ".venv",
  "build",
  "venv",
  "__pycache__",
  "*/migrations/*",
  ".vscode*",
]
line-length = 88


[tool.ruff.lint]
select = [
  "B",      # flake8-bugbear
  "BLE",    # flake8-blind-except
  "C4",     # flake8-comprehensions
  "DJ",     # flake8-django
  "I",      # isort
  "PLC",    # pylint-convention
  "PLE",    # pylint-error
  "PLR",    # pylint-refactoring
  "PLW",    # pylint-warning
  "RUF100", # Ruff unused-noqa
  "RUF200", # Ruff check pyproject.toml
  "S",      # flake8-bandit
  "SLF",    # flake8-self
  "T20",    # flake8-print
]
ignore = ["DJ001", "PLR2004"]

[tool.ruff.lint.isort]
section-order = [
  "future",
  "standard-library",
  "django",
  "third-party",
  "find",
  "first-party",
  "local-folder",
]
sections = { find = ["core"], django = ["django"] }

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["S", "SLF"]

[tool.pytest.ini_options]
addopts = [
  "-v",
  "--cov-report",
  "term-missing",
  # Allow test files to have the same name in different directories.
  "--import-mode=importlib",
]
python_files = ["test_*.py", "tests.py"]
